a bit more repo cleanup
