+++ llm-meditators/webapp/.vscode/settings.json
@@ -13,5 +13,10 @@
       "source.fixAll.eslint": true
     },
     "editor.formatOnSave": true
+  },
+  "[html][typescript]": {
+    "editor.codeActionsOnSave": {
+      "source.fixAll.eslint": "explicit"
+    }
   }
 }
