+++ llm-meditators/webapp/src/app/app-settings/app-settings.component.ts
@@ -7,13 +7,13 @@
 ==============================================================================*/
 
 import { Component, ElementRef, OnInit, ViewChild, effect } from '@angular/core';
-import { AppData, SavedDataService } from '../saved-data.service';
+import { AppData, SavedDataService } from '../services/saved-data.service';
 import { FormControl } from '@angular/forms';
-import { LmApiService } from '../lm-api.service';
-import { GoogleSheetsService, isSheetsError } from '../google-sheets.service';
-import { GoogleAuthService } from '../google-auth.service';
-import { GoogleDriveAppdataService } from '../google-drive-appdata.service';
-import { ItemInterpreterService } from '../item-interpreter.service';
+import { LmApiService } from '../services/lm-api.service';
+import { GoogleSheetsService, isSheetsError } from '../services/google-sheets.service';
+import { GoogleAuthService } from '../services/google-auth.service';
+import { GoogleDriveAppdataService } from '../services/google-drive-appdata.service';
+import { ItemInterpreterService } from '../services/item-interpreter.service';
 import { SimpleError, isErrorResponse } from 'src/lib/simple-errors/simple-errors';
