--- llm-meditators/webapp/src/app/llm-api-config/llm-api-config.component.html
@@ -1,33 +1,0 @@
-<div class="main">
-
-<p>
-You can this on the command line to get your parameters:
-</p>
-<pre>
-echo "Project: $(gcloud config get-value project)"
-echo "AccessToken: $(gcloud auth print-access-token)"
-</pre>
-<p>
-Note: the access token generally works for 1 hour. After that, you'll need to
-renew it.
-</p>
-<p>
-And then copy/paste them here:
-</p>
-<form class="cloud-info-form">
-  <mat-form-field class="full-width"  >
-    <mat-label>Project</mat-label>
-    <input type="text" matInput [formControl]="projectControl">
-  </mat-form-field>
-
-  <mat-form-field class="full-width">
-    <mat-label>AccessToken</mat-label>
-    <textarea matInput [formControl]="tokenControl"></textarea>
-  </mat-form-field>
-</form>
-<p>
-  These values are saved in your local browser storage, you can clear them
-  by simply editing the above fields (local browser storage is updated as you
-  type).
-</p>
-</div>
