--- llm-meditators/webapp/src/app/llm-api-config/llm-api-config.component.scss
@@ -1,12 +1,0 @@
-.main {
-}
-
-.cloud-info-form {
-  min-width: 150px;
-  max-width: 500px;
-  width: 100%;
-}
-
-.full-width {
-  width: 100%;
-}
