--- llm-meditators/webapp/src/app/llm-api-config/llm-api-config.component.spec.ts
@@ -1,50 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an Apache2 license that can be
- * found in the LICENSE file and http://www.apache.org/licenses/LICENSE-2.0
-==============================================================================*/
-
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { LlmApiConfigComponent } from './llm-api-config.component';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatInputModule } from '@angular/material/input';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatListModule } from '@angular/material/list';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-
-describe('LlmApiConfigComponent', () => {
-  let component: LlmApiConfigComponent;
-  let fixture: ComponentFixture<LlmApiConfigComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [LlmApiConfigComponent],
-      imports: [
-        NoopAnimationsModule,
-        MatButtonModule,
-        MatIconModule,
-        MatListModule,
-        FormsModule,
-        ReactiveFormsModule,
-        MatFormFieldModule,
-        MatInputModule,
-        MatMenuModule,
-        MatProgressBarModule,
-      ],
-    });
-    fixture = TestBed.createComponent(LlmApiConfigComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
