--- llm-meditators/webapp/src/app/prompts-config/prompts-config.component.html
@@ -1,1 +1,0 @@
-<p>prompts-config works!</p>
