--- llm-meditators/webapp/src/app/prompts-config/prompts-config.component.ts
@@ -1,18 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an Apache2 license that can be
- * found in the LICENSE file and http://www.apache.org/licenses/LICENSE-2.0
-==============================================================================*/
-
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-prompts-config',
-  templateUrl: './prompts-config.component.html',
-  styleUrls: ['./prompts-config.component.scss']
-})
-export class PromptsConfigComponent {
-
-}
